requirements.txt
Django==1.4
gunicorn==0.14.2
kombu==2.1.7
psycopg2==2.4.5
virtualenv==1.6.4
wsgiref==0.1.2
simplejson
django-tastypie